2d st-dev histogramm

+++ src/godoy/Exporter.java
@@ -658,7 +658,7 @@ public class Exporter {
 	
 	public void exportPeaksPositionHistogramm() {
 		try {
-			int[] histogramm = analyzer.getHistogramm();
+			double[] histogramm = analyzer.getHistogramm();
 			
 			int size = 400;
